requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tenacity>=8.2.0
schedule>=1.2.0
rich>=13.0.0
pydantic>=2.0.0